small fixes and added more docs
