+++ .circleci/smoketest_docker.sh
@@ -110,9 +110,6 @@ fakeversion=`cat VERSION`
 fakeversion="${fakeversion}.ci"
 echo $fakeversion > ./VERSION
 
-# We only build linux/amd64 image since Circle CI machine image has some issues with arm. This is
-# fine since we only test amd64 image on Circle CI.
-#python3 build_package_new.py "${log_mode}-debian" --tag "local_image" --coverage --platforms linux/amd64
 docker image ls
 echo "::endgroup::"
 
@@ -120,10 +117,10 @@ echo "::endgroup::"
 # Launch Agent container (which begins gathering stdout logs)
 echo "::group::Launch Agent container (which begins gathering stdout logs)"
 docker run -d --name ${contname_agent} \
-  -e SCALYR_API_KEY=${SCALYR_API_KEY} -e SCALYR_SERVER=${SCALYR_SERVER} \
-  -v /var/run/docker.sock:/var/scalyr/docker.sock \
-  ${jsonlog_containers_mount} ${syslog_driver_portmap} \
-  "${tested_image_name}"
+-e SCALYR_API_KEY=${SCALYR_API_KEY} -e SCALYR_SERVER=${SCALYR_SERVER} \
+-v /var/run/docker.sock:/var/scalyr/docker.sock \
+${jsonlog_containers_mount} ${syslog_driver_portmap} \
+"${tested_image_name}"
 echo "::endgroup::"
 
 # Capture agent short container ID
