+++ .circleci/smoketest_k8s.sh
@@ -111,11 +111,6 @@ echo "=================================================="
 # Build local image (add .ci.k8s to version)
 perl -pi.bak -e 's/\s*(\S+)/$1\.ci\.k8s/' VERSION
 
-# Build image by specifying image type through build args.
-# We only build linux/amd64 image since Circle CI machine image has some issues with arm. This is
-# fine since we only test amd64 image on Circle CI.
-#python3 build_package_new.py k8s-debian --tag "local_k8s_image" --coverage --platforms linux/amd64
-
 docker image ls
 echo "::endgroup::"
