+++ agent_build_refactored/utils/docker/common.py
@@ -1,3 +1,18 @@
+# Copyright 2014-2023 Scalyr Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
 import json
 import logging
 import subprocess
@@ -11,6 +26,9 @@ def get_docker_container_host_port(
     container_port: str,
     prefix_cmd_args: List[str] = None,
 ):
+    """
+    Get a host-side port number of the container.
+    """
 
     prefix_cmd_args = prefix_cmd_args or []
 
@@ -37,7 +55,6 @@ def get_docker_container_host_port(
     container_info = inspect_result[0]
     host_port = container_info["NetworkSettings"]["Ports"][container_port][0]["HostPort"]
     return host_port
-
 
 
 def delete_container(
