--- tests/end_to_end_tests/container_images_test/test.py
@@ -1,2 +1,0 @@
-def test_d(test_image_tag):
-    pass
